Add is_invocable
